Refactor Android progress surfaces
