+++ apps/android/feature/progress/src/main/java/com/flashcardsopensourceapp/feature/progress/ProgressSectionDefaults.kt
@@ -1,0 +1,12 @@
+package com.flashcardsopensourceapp.feature.progress
+
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.ui.unit.dp
+
+internal val progressSectionShape = RoundedCornerShape(28.dp)
+
+const val progressStreakSectionTag: String = "progress_streak_section"
+const val progressReviewsSectionTag: String = "progress_reviews_section"
+const val progressReviewsActivityChartTag: String = "progress_reviews_activity_chart"
+const val progressReviewScheduleSectionTag: String = "progress_review_schedule_section"
+const val progressReviewScheduleDonutChartTag: String = "progress_review_schedule_donut_chart"
